local_requirements.txt
# Core requirements for local testing
torch>=2.0.0
pytorch-lightning>=2.0.0
torchmetrics>=0.11.0
einops>=0.4.0
loguru>=0.6.0
pyyaml>=6.0
matplotlib>=3.5.0
numpy>=1.20.0
markupsafe==2.0.1
packaging>=21.0
fvcore>=0.1.5
seaborn>=0.12
transformers==4.28.0
torchvision>=0.15
opencv-python>=4.6
pandas>=1.4

# Required for testing perceiver-io
datasets>=2.4.0
tensorboard>=2.10.0
